# reduce inference and optimization time (by turning it off)