@KristofferC 2021-03-11
Would be nice with a new release whenever is suitable.