@oscardssmith 2021-03-11
Is there any other low hanging fruit for speeding this up (better precompile or something?)