@KristofferC 2021-03-11
You could experiment with SnoopCompile.jl.